pyproject.toml
[project]
name = "stockanalyse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "akshare",
    "pandas",
    "openpyxl",
    "python-dotenv",
    "jieba==0.42.1",
    "pypdf==5.8.0",
    "Requests==2.32.3",
    "xlwt==1.3.0",
    "langchain==0.3.26",
    "langchain-openai==0.3.28",
    "langchain-community==0.3.27",
    "langchain-huggingface",
    "langchain-chroma",
    "streamlit==1.47.0",
    "wikipedia",
    "sentence-transformers",
    "unstructured[csv,md,pdf]>=0.18.11",
    "pdfminer-six==20240706",
    "py-mini-racer>=0.6.0",
]
